Split some APIs from nn.py to loss.py
test=develop
